+++ src/Utility/Sitemap/SitemapBuilder.php
@@ -28,15 +28,14 @@ class SitemapBuilder
     /**
      * Gets all executable methods for the `Sitemap` class of a plugin
      * @param string $plugin Plugin
-     * @return \Cake\Collection\CollectionInterface<array{class: class-string, name: string}> Collection of classes and methods names
+     * @return array<array{class: class-string, name: string}> Array with classes and methods names
      * @throws \ErrorException
      */
-    public static function getMethods(string $plugin): CollectionInterface
+    public static function getMethods(string $plugin): array
     {
         $class = $plugin . '\Utility\Sitemap\Sitemap';
 
-        return collection(class_exists($class) ? get_child_methods($class) : [])
-            ->map(fn(string $name): array => ['class' => $class] + compact('name'));
+        return class_exists($class) ? array_map(fn(string $name) => ['class' => $class] + compact('name'), get_child_methods($class)) : [];
     }
 
     /**